requirements.txt
# Core
Django>=5.0,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.3,<4.5

# Server
gunicorn>=21.2,<22
whitenoise==6.11.0

# Google Gemini (PIN THIS)
google-genai>=0.3.0,<0.4.0
pydantic>=2.6,<2.9

# DB
psycopg2-binary>=2.9,<3.0

# AWS
boto3>=1.28,<2.0

# Utils
requests>=2.31,<3.0
pypdf>=3.11,<4.0
python-docx>=0.8.11,<1.0
pillow>=10.0,<11.0
pytesseract>=0.3.13,<0.4.0

# OpenAI
openai>=1.0.0,<2.0
